requirements.txt
kafka-python>=2.0.2
pandas>=2.2.0
cassandra-driver>=3.29.0
vnstock>=0.2.8.8
mysql-connector-python>=8.0.36
SQLAlchemy>=2.0.25
bs4>=0.0.2
ssi_fc_data>=2.2.1
pyspark~=3.5.0
pyarrow>=16.0.0
pip~=22.0.2
distro~=1.7.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.0.1
setuptools~=59.6.0
typing_extensions~=4.11.0
zipp~=1.0.0
future~=0.18.2